[ZMRT-4] Fix a bug in route construction
There was a bug in route construction: for a line transfer, the
transfer station should have two station codes in the path, one for the
from-line and another for the to-line. The station code for the to-line
was omitted. This is now fixed.

The logic for route construction has been moved into a separate class,
and unit tests have also been included.